Only call Navision API on external fulfillment
Calls to Navision API should only happen on external, not local, fulfillment. Also set order status to ORDERED so that shipment can be created (system only allows shipments to be created for orders in ORDERED or FULFILLING status). Set an extra data property of externallyFulfilled to true so that when the shipment is created, it can be checked to bypass a transfer stock event processed during the default processor.